refactor: 패키지 변경 - enums 패키지 삭제 후 domain 패키지로 이동
