+++ src/main/kotlin/blackjack/domain/CardTier.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.enums
+package blackjack.domain
 
 enum class CardTier(
     val values: Int,
