+++ src/main/kotlin/blackjack/domain/TrumpCard.kt
@@ -1,8 +1,5 @@
 package blackjack.domain
 
-import blackjack.domain.enums.CardTier
-import blackjack.domain.enums.Shape
-
 data class TrumpCard(
     val tier: CardTier,
     val shape: Shape,
